requirements.txt
flask==3.0.3
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.32.3
openpyxl==3.1.5
pandas>=1.5.0
pyecharts>=2.0.0
